# openapi.yaml
openapi: 3.0.3
info:
  title: 잇올 1:1 멘토링 플랫폼 API
  description: >
    선생님–학생 1:1 멘토링/상담 예약 플랫폼 REST API (초안).
    근거: 통합 스펙 v2.5 · ERD-API 초안 · 스택 NestJS+PostgreSQL.
    민감정보(인적사항·상담내용·결제·급여)는 권한별 마스킹.
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/member-types:
    get:
      tags: [HR & People]
      summary: 회원 분류 목록(선생님·학생 유형) — 관리자/HR
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [teacher, student] } }
      responses: { '200': { description: ok } }
    post:
      tags: [HR & People]
      summary: 회원 분류 추가(본사 이상 L2) — 확장 가능
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, code, label]
              properties:
                kind: { type: string, enum: [teacher, student] }
                code: { type: string }
                label: { type: string }
                sortOrder: { type: integer }
      responses:
        '201': { description: created }
        '403': { $ref: '#/components/responses/Error' }
  /admin/member-types/{id}:
    delete:
      tags: [HR & People]
      summary: 회원 분류 비활성(본사 이상 L2)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: ok }, '403': { $ref: '#/components/responses/Error' } }
  /admin/teachers/{id}/type:
    patch:
      tags: [HR & People]
      summary: 선생님 분류 배정(관리자/HR)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [typeCode], properties: { typeCode: { type: string } } }
      responses: { '200': { description: ok } }
  /admin/students/{id}/type:
    patch:
      tags: [HR & People]
      summary: 학생 분류 배정(관리자/HR)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [typeCode], properties: { typeCode: { type: string } } }
      responses: { '200': { description: ok } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...